requirements.txt
# Web Crawling & HTTP
aiohttp>=3.8.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
requests>=2.28.0
httpx>=0.24.0

# Pattern Matching & Analysis
pyyaml>=5.4.0
python-dotenv>=0.19.0

# AI Integration (Multiple Providers)
openai>=1.0.0
google-generativeai>=0.3.0
ollama>=0.1.0

# Cloud SDK Validators
boto3==1.34.10
stripe==7.8.0
PyGithub==2.1.1
twilio==8.11.1

# Caching & Performance
diskcache==5.6.3

# CLI & Output
click==8.1.7
colorama==0.4.6
tqdm>=4.65.0
tqdm==4.66.1
tabulate==0.9.0

# Utilities
urllib3==2.1.0
chardet==5.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
